fix(test): 4468 회귀 unique 쪽 번호 타입을 u32로 맞춘다

+++ tests/cases/issue_4468_split_cell_float_once.rs
@@ -92,7 +92,7 @@ fn issue_4468_split_cell_tb_float_paints_once() {
         "셀 TopAndBottom 그림이 한 장도 없다"
     );
     for (id, pages) in &pages_by_id {
-        let unique: Vec<usize> = {
+        let unique: Vec<u32> = {
             let mut p = pages.clone();
             p.sort_unstable();
             p.dedup();
